requirements.txt
apache-airflow
pandas
requests
beautifulsoup4
lxml